Move some more types into core.
